use uint32_t instead of uint8_t

+++ src/ic_vec.c
@@ -90,8 +90,8 @@ vector ic_reads_vec_int64(vector v){
   match_byte(buf, len, 6, IDL_TYPE_INT64);
   match_byte(buf, len, 7, 0x01);
   match_byte(buf, len, 8, 0x00);
-  uint8_t budget_var=1;
-  uint8_t *budget;
+  uint32_t budget_var=1;
+  uint32_t *budget;
   budget = &budget_var;
   uint32_t vec_length= read_uleb128(buf,9,budget);
   int64_t result = 0;
@@ -135,8 +135,8 @@ vector ic_reads_vec_nat64(vector v){
   match_byte(buf, len, 6, IDL_TYPE_NAT64);
   match_byte(buf, len, 7, 0x01);
   match_byte(buf, len, 8, 0x00);
-  uint8_t budget_var=1;
-  uint8_t *budget;
+  uint32_t budget_var=1;
+  uint32_t *budget;
   budget = &budget_var;
   uint32_t vec_length= read_uleb128(buf,9,budget);
   uint64_t result = 0;
@@ -180,8 +180,8 @@ vector ic_reads_vec_int32(vector v){
   match_byte(buf, len, 6, IDL_TYPE_INT32);
   match_byte(buf, len, 7, 0x01);
   match_byte(buf, len, 8, 0x00);
-  uint8_t budget_var=1;
-  uint8_t *budget;
+  uint32_t budget_var=1;
+  uint32_t *budget;
   budget = &budget_var;
   uint32_t vec_length= read_uleb128(buf,9,budget);
   int32_t result = 0;
@@ -225,8 +225,8 @@ vector ic_reads_vec_nat32(vector v){
   match_byte(buf, len, 6, IDL_TYPE_NAT32);
   match_byte(buf, len, 7, 0x01);
   match_byte(buf, len, 8, 0x00);
-  uint8_t budget_var=1;
-  uint8_t *budget;
+  uint32_t budget_var=1;
+  uint32_t *budget;
   budget = &budget_var;
   uint32_t vec_length= read_uleb128(buf,9,budget);
   uint32_t result = 0;
@@ -270,8 +270,8 @@ vector ic_reads_vec_int16(vector v){
   match_byte(buf, len, 6, IDL_TYPE_INT16);
   match_byte(buf, len, 7, 0x01);
   match_byte(buf, len, 8, 0x00);
-  uint8_t budget_var=1;
-  uint8_t *budget;
+  uint32_t budget_var=1;
+  uint32_t *budget;
   budget = &budget_var;
   uint32_t vec_length= read_uleb128(buf,9,budget);
   int16_t result = 0;
@@ -316,8 +316,8 @@ vector ic_reads_vec_nat16(vector v){
   match_byte(buf, len, 6, IDL_TYPE_NAT16);
   match_byte(buf, len, 7, 0x01);
   match_byte(buf, len, 8, 0x00);
-  uint8_t budget_var=1;
-  uint8_t *budget;
+  uint32_t budget_var=1;
+  uint32_t *budget;
   budget = &budget_var;
   uint32_t vec_length= read_uleb128(buf,9,budget);
   uint16_t result = 0;
@@ -363,8 +363,8 @@ vector ic_reads_vec_int8(vector v){
   match_byte(buf, len, 6, IDL_TYPE_INT8);
   match_byte(buf, len, 7, 0x01);
   match_byte(buf, len, 8, 0x00);
-  uint8_t budget_var=1;
-  uint8_t *budget;
+  uint32_t budget_var=1;
+  uint32_t *budget;
   budget = &budget_var;
   uint32_t vec_length= read_uleb128(buf,9,budget);
   int8_t result = 0;
@@ -409,8 +409,8 @@ vector ic_reads_vec_nat8(vector v){
   match_byte(buf, len, 6, IDL_TYPE_NAT8);
   match_byte(buf, len, 7, 0x01);
   match_byte(buf, len, 8, 0x00);
-  uint8_t budget_var=1;
-  uint8_t *budget;
+  uint32_t budget_var=1;
+  uint32_t *budget;
   budget = &budget_var;
   uint32_t vec_length= read_uleb128(buf,9,budget);
   uint8_t result = 0;
